Add interning for futurers
