+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/types/BFutureType.java
@@ -39,6 +39,7 @@ import java.util.Set;
 public class BFutureType extends BType implements FutureType {
 
     private static final BasicTypeBitSet BASIC_TYPE = Builder.getFutureType();
+    private static final SimpleTypeCheckFlyweightStore FLYWEIGHT_CACHE = new SimpleTypeCheckFlyweightStore();
     private final Type constraint;
 
     /**
@@ -53,8 +54,11 @@ public class BFutureType extends BType implements FutureType {
     }
 
     public BFutureType(Type constraint) {
-        super(TypeConstants.FUTURE_TNAME, null, Object.class, true);
+        super(TypeConstants.FUTURE_TNAME, null, Object.class, false);
         this.constraint = constraint;
+        var flyweight = FLYWEIGHT_CACHE.get(constraint);
+        this.typeCheckCache = flyweight.typeCheckCache();
+        this.typeId = flyweight.typeId();
     }
 
     public Type getConstrainedType() {
