+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/internal/types/BTypedescType.java
@@ -18,7 +18,6 @@
 
 package io.ballerina.runtime.internal.types;
 
-import com.github.benmanes.caffeine.cache.LoadingCache;
 import io.ballerina.runtime.api.Module;
 import io.ballerina.runtime.api.constants.TypeConstants;
 import io.ballerina.runtime.api.types.PredefinedTypes;
@@ -27,17 +26,12 @@ import io.ballerina.runtime.api.types.TypeTags;
 import io.ballerina.runtime.api.types.TypedescType;
 import io.ballerina.runtime.api.types.semtype.BasicTypeBitSet;
 import io.ballerina.runtime.api.types.semtype.Builder;
-import io.ballerina.runtime.api.types.semtype.CacheableTypeDescriptor;
 import io.ballerina.runtime.api.types.semtype.Context;
 import io.ballerina.runtime.api.types.semtype.SemType;
-import io.ballerina.runtime.api.types.semtype.TypeCheckCache;
-import io.ballerina.runtime.api.types.semtype.TypeCheckCacheFactory;
-import io.ballerina.runtime.internal.types.semtype.CacheFactory;
 import io.ballerina.runtime.internal.types.semtype.TypedescUtils;
 import io.ballerina.runtime.internal.values.TypedescValue;
 import io.ballerina.runtime.internal.values.TypedescValueImpl;
 
-import java.util.Map;
 import java.util.Set;
 
 /**
@@ -48,6 +42,7 @@ import java.util.Set;
 public class BTypedescType extends BType implements TypedescType {
 
     private static final BasicTypeBitSet BASIC_TYPE = Builder.getTypeDescType();
+    private static final SimpleTypeCheckFlyweightStore FLYWEIGHT_CACHE = new SimpleTypeCheckFlyweightStore();
 
     private final Type constraint;
 
@@ -59,7 +54,7 @@ public class BTypedescType extends BType implements TypedescType {
     public BTypedescType(Type constraint) {
         super(TypeConstants.TYPEDESC_TNAME, null, TypedescValue.class, false);
         this.constraint = constraint;
-        var flyweight = TypeCheckFlyweightStore.get(constraint);
+        var flyweight = FLYWEIGHT_CACHE.get(constraint);
         this.typeCheckCache = flyweight.typeCheckCache();
         this.typeId = flyweight.typeId();
     }
@@ -125,59 +120,4 @@ public class BTypedescType extends BType implements TypedescType {
                 constraintType.isDependentlyTyped(visited);
     }
 
-    private static class TypeCheckFlyweightStore {
-
-        private static final LoadingCache<Integer, TypeCheckFlyweight> unnamedTypeCache =
-                CacheFactory.createCache(TypeCheckFlyweight::new);
-
-        private static final TypeCheckFlyweight[] reservedLAT =
-                new TypeCheckFlyweight[TypeIdSupplier.MAX_RESERVED_ID];
-
-        private static final Map<Integer, TypeCheckFlyweight> namedTypeCache = CacheFactory.createCachingHashMap();
-
-        public static TypeCheckFlyweight get(Type constraint) {
-            if (constraint instanceof CacheableTypeDescriptor cacheableTypeDescriptor) {
-                return getInner(cacheableTypeDescriptor.typeId());
-            }
-            return new TypeCheckFlyweight(TypeIdSupplier.getAnonId(), TypeCheckCacheFactory.create());
-        }
-
-        private static TypeCheckFlyweight getInner(int typeId) {
-            return switch (TypeIdSupplier.kind(typeId)) {
-                case RESERVED -> getReserved(typeId);
-                case NAMED -> getNamed(typeId);
-                case UNNAMED -> getUnnamed(typeId);
-            };
-        }
-
-        private static TypeCheckFlyweight getUnnamed(int typeId) {
-            return unnamedTypeCache.get(typeId);
-        }
-
-        private static TypeCheckFlyweight getNamed(int typeId) {
-            var cached = namedTypeCache.get(typeId);
-            if (cached != null) {
-                return cached;
-            }
-            cached = new TypeCheckFlyweight(typeId);
-            namedTypeCache.put(typeId, cached);
-            return cached;
-        }
-
-        private static TypeCheckFlyweight getReserved(int typeId) {
-            TypeCheckFlyweight o = reservedLAT[typeId];
-            if (o == null) {
-                o = new TypeCheckFlyweight(typeId);
-                reservedLAT[typeId] = o;
-            }
-            return o;
-        }
-
-        private record TypeCheckFlyweight(int typeId, TypeCheckCache typeCheckCache) {
-
-            public TypeCheckFlyweight(Integer constraintId) {
-                this(TypeIdSupplier.getAnonId(), TypeCheckCacheFactory.create());
-            }
-        }
-    }
 }
